+++ src/archive/lendingcontract.sol
@@ -6,7 +6,7 @@ import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 import "@openzeppelin/contracts/utils/ReentrancyGuard.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 import "@openzeppelin/contracts/utils/cryptography/ECDSA.sol";
-import { console2 } from "forge-std/src/Test.sol"; // TODO REMOVE AFDTRER TEST
+import {console2} from "forge-std/src/Test.sol"; // TODO REMOVE AFDTRER TEST
 
 interface IWETH {
     function deposit() external payable;
@@ -26,10 +26,10 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
         uint256 timestamp;
         uint256 rate;
     }
-    
+
     address public issuer;
     IWETH public immutable WETH;
-    
+
     InterestRateChange[] public interestRateHistory;
     mapping(uint256 => mapping(address => BorrowPosition)) public borrowPositions; // nftId => wallet => BorrowPosition
     mapping(uint256 => uint256) public borrowNonces; // nftId => nonce
@@ -55,7 +55,7 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
 
     event AutogasTriggered(uint256 indexed nftId, address indexed wallet, uint256 amount);
     event AutogasSpikeTriggered(uint256 indexed nftId, address indexed wallet, uint256 amount);
-    
+
     event PoolDeposited(uint256 amount);
     event PoolWithdrawn(uint256 amount);
 
@@ -64,16 +64,15 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
     event InterestRateChanged(uint256 oldRate, uint256 newRate);
     event GasPriceThresholdChanged(uint256 newThreshold);
 
-
     constructor(address _issuer, address _weth, uint256 _chainId) Ownable(msg.sender) {
         issuer = _issuer;
         feeRecipient = _issuer;
         WETH = IWETH(_weth);
         borrowInterestRate = 1000; // 10% annual interest
         autogasThreshold = 1e15; // 0.001 ETH
-        autogasRefillAmount = 5e14  ; // 0.0005 ETH, adjust as needed
+        autogasRefillAmount = 5e14; // 0.0005 ETH, adjust as needed
         repaymentThreshold = 1e13; // 0.00001 ETH aka $0.035 at current prices
-        gasPriceThreshold = 2; 
+        gasPriceThreshold = 2;
         chainId = _chainId;
     }
 
@@ -86,13 +85,10 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
         _;
     }
 
-    function borrow(
-        uint256 nftId,
-        uint256 amount,
-        uint256 timestamp,
-        uint256 nonce,
-        bytes memory signature
-    ) external nonReentrant {
+    function borrow(uint256 nftId, uint256 amount, uint256 timestamp, uint256 nonce, bytes memory signature)
+        external
+        nonReentrant
+    {
         require(amount > 0, "Borrow amount must be greater than 0");
         require(block.timestamp <= timestamp + SIGNATURE_VALIDITY, "Signature expired");
         require(nonce == borrowNonces[nftId], "Invalid nonce");
@@ -117,16 +113,16 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
         require(amount > 0, "Borrow amount must be greater than 0");
         require(block.timestamp <= timestamp + SIGNATURE_VALIDITY, "Signature expired");
         require(nonce == borrowNonces[nftId], "Invalid nonce");
-        
-        // The borrower can be different from the signer. The signer signs with borrower's address in the signature. 
-        // The issuer verifies that the signature is from the signer himself. If not he won't approve it. 
+
+        // The borrower can be different from the signer. The signer signs with borrower's address in the signature.
+        // The issuer verifies that the signature is from the signer himself. If not he won't approve it.
         bytes32 messageHash = keccak256(abi.encodePacked(recipient, nftId, amount, timestamp, nonce, chainId));
         bytes32 ethSignedMessageHash = getEthSignedMessageHash(messageHash);
-        
+
         address signer = ethSignedMessageHash.recover(userSignature);
         require(ethSignedMessageHash.recover(issuerSignature) == issuer, "Invalid signature");
 
-        console2.log('recovered signer', signer);
+        console2.log("recovered signer", signer);
 
         uint256 gasStart = gasleft();
         _executeBorrowAndSend(nftId, signer, amount, recipient);
@@ -138,9 +134,10 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
 
     function _executeBorrow(uint256 nftId, address wallet, uint256 amount) internal {
         BorrowPosition storage borrowPosition = borrowPositions[nftId][wallet];
-        
-        uint256 borrowInterest = calculateCompoundInterest(borrowPosition.amount, borrowPosition.timestamp, borrowInterestRate);
-        
+
+        uint256 borrowInterest =
+            calculateCompoundInterest(borrowPosition.amount, borrowPosition.timestamp, borrowInterestRate);
+
         if (!isBorrower(nftId, wallet)) {
             borrowers[nftId].push(wallet);
         }
@@ -148,11 +145,11 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
         borrowPosition.amount += borrowInterest + amount + borrowFees;
         borrowPosition.timestamp = block.timestamp;
 
-        WETH.withdraw(amount+borrowFees);
-        (bool success, ) = wallet.call{value: amount}("");
+        WETH.withdraw(amount + borrowFees);
+        (bool success,) = wallet.call{value: amount}("");
         require(success, "ETH transfer failed");
-        if (borrowFees>0) {
-            (success, ) = feeRecipient.call{value: borrowFees}("");
+        if (borrowFees > 0) {
+            (success,) = feeRecipient.call{value: borrowFees}("");
             require(success, "ETH transfer failed");
         }
 
@@ -161,9 +158,10 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
 
     function _executeBorrowAndSend(uint256 nftId, address wallet, uint256 amount, address recipient) internal {
         BorrowPosition storage borrowPosition = borrowPositions[nftId][wallet];
-        
-        uint256 borrowInterest = calculateCompoundInterest(borrowPosition.amount, borrowPosition.timestamp, borrowInterestRate);
-        
+
+        uint256 borrowInterest =
+            calculateCompoundInterest(borrowPosition.amount, borrowPosition.timestamp, borrowInterestRate);
+
         if (!isBorrower(nftId, wallet)) {
             borrowers[nftId].push(wallet);
         }
@@ -171,11 +169,11 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
         borrowPosition.amount += borrowInterest + amount + borrowFees;
         borrowPosition.timestamp = block.timestamp;
 
-        WETH.withdraw(amount+borrowFees);
-        (bool success, ) = recipient.call{value: amount}("");
+        WETH.withdraw(amount + borrowFees);
+        (bool success,) = recipient.call{value: amount}("");
         require(success, "ETH transfer failed");
-        if (borrowFees>0) {
-            (success, ) = feeRecipient.call{value: borrowFees}("");
+        if (borrowFees > 0) {
+            (success,) = feeRecipient.call{value: borrowFees}("");
             require(success, "ETH transfer failed");
         }
 
@@ -184,7 +182,7 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
 
     function isBorrower(uint256 nftId, address wallet) internal view returns (bool) {
         address[] storage nftBorrowers = borrowers[nftId];
-        for (uint i = 0; i < nftBorrowers.length; i++) {
+        for (uint256 i = 0; i < nftBorrowers.length; i++) {
             if (nftBorrowers[i] == wallet) {
                 return true;
             }
@@ -215,82 +213,100 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
 
     function repay(uint256 nftId, address wallet, address refundAddress) external payable nonReentrant {
         require(msg.value > 0, "Repay amount must be greater than 0");
-        
+
         BorrowPosition storage borrowPosition = borrowPositions[nftId][wallet];
         require(borrowPosition.amount > 0, "No borrow position for this NFT and wallet");
 
-        uint256 borrowInterest = calculateCompoundInterest(borrowPosition.amount, borrowPosition.timestamp, borrowInterestRate);
+        uint256 borrowInterest =
+            calculateCompoundInterest(borrowPosition.amount, borrowPosition.timestamp, borrowInterestRate);
         uint256 totalOwed = borrowPosition.amount + borrowInterest;
-        
+
         uint256 repayAmount = msg.value > totalOwed ? totalOwed : msg.value;
         borrowPosition.amount = totalOwed > repayAmount ? totalOwed - repayAmount : 0;
         borrowPosition.timestamp = block.timestamp;
 
         WETH.deposit{value: repayAmount}();
-        
+
         if (repayAmount < msg.value) {
             uint256 refundAmount = msg.value - repayAmount;
-            (bool success, ) = refundAddress.call{value: refundAmount}("");
+            (bool success,) = refundAddress.call{value: refundAmount}("");
             require(success, "ETH refund failed");
         }
 
         emit Repaid(nftId, wallet, repayAmount);
     }
 
-    function repayMultiple(uint256[] memory nftIds, address[] memory wallets, uint256[] memory amounts, address refundAddress) external payable nonReentrant {
+    function repayMultiple(
+        uint256[] memory nftIds,
+        address[] memory wallets,
+        uint256[] memory amounts,
+        address refundAddress
+    ) external payable nonReentrant {
         require(nftIds.length == wallets.length && wallets.length == amounts.length, "Arrays length mismatch");
         require(msg.value > 0, "Repay amount must be greater than 0");
-        
+
         uint256 totalRepaid = 0;
-        
-        for (uint i = 0; i < nftIds.length; i++) {
+
+        for (uint256 i = 0; i < nftIds.length; i++) {
             BorrowPosition storage borrowPosition = borrowPositions[nftIds[i]][wallets[i]];
             require(borrowPosition.amount > 0, "No borrow position for this NFT and wallet");
-    
-            uint256 borrowInterest = calculateCompoundInterest(borrowPosition.amount, borrowPosition.timestamp, borrowInterestRate);
+
+            uint256 borrowInterest =
+                calculateCompoundInterest(borrowPosition.amount, borrowPosition.timestamp, borrowInterestRate);
             uint256 totalOwed = borrowPosition.amount + borrowInterest;
-            
+
             uint256 repayAmount = amounts[i] > totalOwed ? totalOwed : amounts[i];
             borrowPosition.amount = totalOwed - repayAmount;
             borrowPosition.timestamp = block.timestamp;
-    
+
             totalRepaid += repayAmount;
             emit Repaid(nftIds[i], wallets[i], repayAmount);
         }
-    
+
         require(totalRepaid <= msg.value, "Insufficient ETH sent");
         WETH.deposit{value: totalRepaid}();
-        
+
         if (totalRepaid < msg.value) {
-            (bool success, ) = refundAddress.call{value: msg.value - totalRepaid}("");
+            (bool success,) = refundAddress.call{value: msg.value - totalRepaid}("");
             require(success, "ETH refund failed");
         }
     }
 
-    function calculateCompoundInterest(uint256 principal, uint256 lastUpdateTime, uint256 currentInterestRate) internal view returns (uint256) {
+    function calculateCompoundInterest(uint256 principal, uint256 lastUpdateTime, uint256 currentInterestRate)
+        internal
+        view
+        returns (uint256)
+    {
         if (principal == 0) return 0;
-    
+
         uint256 totalInterest = 0;
         uint256 currentPrincipal = principal;
         uint256 currentTime = lastUpdateTime;
-    
-        for (uint i = 0; i < interestRateHistory.length && interestRateHistory[i].timestamp <= block.timestamp; i++) {
+
+        for (uint256 i = 0; i < interestRateHistory.length && interestRateHistory[i].timestamp <= block.timestamp; i++)
+        {
             if (interestRateHistory[i].timestamp > lastUpdateTime) {
-                uint256 periodInterest = calculatePeriodInterest(currentPrincipal, currentTime, interestRateHistory[i].timestamp, currentInterestRate);
+                uint256 periodInterest = calculatePeriodInterest(
+                    currentPrincipal, currentTime, interestRateHistory[i].timestamp, currentInterestRate
+                );
                 totalInterest += periodInterest;
                 currentPrincipal += periodInterest;
                 currentTime = interestRateHistory[i].timestamp;
                 currentInterestRate = interestRateHistory[i].rate;
             }
         }
-    
+
         // Calculate interest for the final period
         totalInterest += calculatePeriodInterest(currentPrincipal, currentTime, block.timestamp, currentInterestRate);
-    
+
         return totalInterest;
     }
-    
-    function calculatePeriodInterest(uint256 principal, uint256 startTime, uint256 endTime, uint256 interestRate) internal pure returns (uint256) {
+
+    function calculatePeriodInterest(uint256 principal, uint256 startTime, uint256 endTime, uint256 interestRate)
+        internal
+        pure
+        returns (uint256)
+    {
         uint256 timeElapsed = endTime - startTime;
         uint256 ratePerSecond = interestRate * 1e18 / (SECONDS_PER_YEAR * 10000); // Convert basis points to per-second rate
         uint256 compoundFactor = compoundExponent(ratePerSecond, timeElapsed);
@@ -302,7 +318,7 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
         uint256 result = 1e18;
         uint256 base = 1e18 + rate;
         while (time > 0) {
-            if (time % 2 == 1) {    
+            if (time % 2 == 1) {
                 result = (result * base) / 1e18;
             }
             base = (base * base) / 1e18;
@@ -313,8 +329,9 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
 
     function getNetPosition(uint256 nftId, address wallet) external view returns (int256) {
         BorrowPosition memory borrowPosition = borrowPositions[nftId][wallet];
-        uint256 borrowAmount = borrowPosition.amount + calculateCompoundInterest(borrowPosition.amount, borrowPosition.timestamp, borrowInterestRate);
-        
+        uint256 borrowAmount = borrowPosition.amount
+            + calculateCompoundInterest(borrowPosition.amount, borrowPosition.timestamp, borrowInterestRate);
+
         if (borrowAmount < repaymentThreshold) {
             return 0; // Consider the debt fully repaid if within the threshold
         }
@@ -323,10 +340,15 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
 
     function getBorrowPosition(uint256 nftId, address wallet) external view returns (uint256 borrowAmount) {
         BorrowPosition memory borrowPosition = borrowPositions[nftId][wallet];
-        borrowAmount = borrowPosition.amount + calculateCompoundInterest(borrowPosition.amount, borrowPosition.timestamp, borrowInterestRate);
+        borrowAmount = borrowPosition.amount
+            + calculateCompoundInterest(borrowPosition.amount, borrowPosition.timestamp, borrowInterestRate);
     }
 
-    function getBorrowPositionSeparate(uint256 nftId, address wallet) external view returns (uint256 borrowAmount, uint256 interestAmount, uint256 borrowTimestamp) {
+    function getBorrowPositionSeparate(uint256 nftId, address wallet)
+        external
+        view
+        returns (uint256 borrowAmount, uint256 interestAmount, uint256 borrowTimestamp)
+    {
         BorrowPosition memory borrowPosition = borrowPositions[nftId][wallet];
         borrowAmount = borrowPosition.amount;
         borrowTimestamp = borrowPosition.timestamp;
@@ -341,7 +363,7 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
 
     function poolWithdraw(uint256 amount) external onlyIssuer {
         WETH.withdraw(amount);
-        (bool success, ) = issuer.call{value: amount}("");
+        (bool success,) = issuer.call{value: amount}("");
         require(success, "ETH transfer failed");
         emit PoolWithdrawn(amount);
     }
@@ -359,7 +381,7 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
         emit InterestRateChangeScheduled(oldRate, newRate, block.timestamp + RATE_CHANGE_DELAY);
     }
 
-    function setInterestRate() external onlyIssuer{
+    function setInterestRate() external onlyIssuer {
         require(scheduleInterestRate != 0, "No scheduled rate change");
         require(block.timestamp > scheduleInterestRateTimestamp + RATE_CHANGE_DELAY, "Delay period not yet passed");
         uint256 oldRate = borrowInterestRate;
@@ -396,7 +418,7 @@ contract CrossChainLendingContract is ReentrancyGuard, Ownable {
         issuer = newIssuer;
     }
 
-    function getBorrowFees() external view returns(uint256) {
+    function getBorrowFees() external view returns (uint256) {
         return borrowFees;
     }
